requirements.txt
mysql-connector-python
requests
opencv-python-headless
numpy
tifffile
pillow
colormath

# TCG scanner additions
fastapi
uvicorn[standard]
python-multipart
imagehash
scipy
pytest